requirements.txt
django==5.1
djangorestframework==3.15
neo4j==5.19
faiss-cpu==1.8.0
torch==2.3.0
torch-geometric==2.5.3
scikit-learn==1.5.0
numpy==1.26.4
redis==5.0
python-dotenv==1.0.1
gunicorn==22.0
psycopg2-binary==2.9.9
